msm: ADSPRPC: optimize logic to set DSP_ROOT
Current logic expects M as a relative path to kernel source and then
get the location of the source code path via $(KERNEL_SRC)/$(M).

But for LE, INPLACE_COMPILE will be set when calling build_module.sh
and M is set to absolute path.

Update logic to set source code path to makefile's path, so that it
can work in both LA and LE case.

Change-Id: I05e182786b33f3fc6d7119102b757139c3dcac5e