dsp-kernel: donate more memory to grow root rpcheap on dsp
Currently, the root rpcheap on dsp is initialized with a 1MB block from
dsp image memory. To reduce the dsp image size, the root rpcheap will
now be initialized with a 512KB block. To compensate for this, the hlos
will donate 1.25MB of memory in the first 3 userpd's (instead of the
1MB that is being donated currently).

Change-Id: I32c302a58c86a35ca00f3506c7fdb0659d27376d
Signed-off-by: Thyagarajan Venkatanarayanan <quic_venkatan@quicinc.com>