dsp-kernel: Classify glink wakeup type on fastrpc channel
When the CPU is in the suspend state,
the DSP may send a signal to the CPU through glink,
and the wakeup will be triggered through glink.
Especially, LPAIDSP will have such usecases.

With this change will classify glink wakeup type on fastrpc channel.
It is helpful for analyzing power-related issues.

Change-Id: Icd14d47b2ed994ebd7674319d88c2dc79d26505e
Signed-off-by: Yongsheng Li <quic_yon@quicinc.com>