dsp-kernel: stop process resource cleanup if the device is about to crash
The current code cleans up resources during device release, even when
the device is on the verge of crashing. This can make debugging difficult
if the same process encounters an exception at the DSP, as resource
cleanup at HLOS obstructs further investigation. To avoid this issue, stop
resource cleanup when the device is about to crash.

Change-Id: I89befb0b517cee4852233dd4f4887599f50df2ee
Signed-off-by: Ramesh Nallagopu <quic_rnallago@quicinc.com>