dsp-kernel: Do not update the status back to the default
After a process is created on the DSP, if it fails to create an
instance of the device due to an allocation failure at the kernel
level, the process state is reset to the default state. This
behavior can potentially be exploited by an attacker to trigger
use-after-free vulnerabilities. To mitigate this risk, the process
status should not be reset to the default state if the PD has been
successfully created on the DSP.

Change-Id: Ib2bdb1bb0c36c6fcca0898c97f97af87cf2b4fc8
Signed-off-by: ssakore <quic_ssakore@quicinc.com>
(cherry picked from commit 80513d510c4e70ba0d3523f2c5390c08ee17f15e)
(cherry picked from commit c4c05e9739752ba9df586869f89eca375251f7bc)