dsp-kernel: Remove return value check for timer_delete_sync
If the invoke call for PD spawn exceeds the timeout value, the timer
callback triggers and queues a work item on the worker thread to
initiate SSR. If the worker thread does not trigger SSR and
timer_delete_sync is called, it returns 0, indicating the timer has
already expired. This return value is incorrectly treated as an error,
causing the invoke call for PD spawn to fail and leading to removal
of mappings from HLOS, which results in an SMMU fault. Remove the return
value check to avoid misinterpreting the timer state and prevent
unnecessary failures.

Change-Id: I4b73864e141ef054c5236002a421b1b06f3ce7da
Signed-off-by: Abhishek Singh <quic_abhishes@quicinc.com>