Adding wild card.
Enabling the compilation of the HY11, HY22 and HWASAN only if
the fake pack script is already run. This is conformed by checking
the adsprpc folder present or not. The adsprpc folder will no be
present after running the fakepack script. This is because the x-ship
rule of adsprpc project is "none" in manifest.xml. Therefore this
folder will get deleted.

Committer: Krishna Dogney <kdogney@quicinc.com>
Change-Id: I7e7c5333a63abdc502a302cf5d71c75274a91003