dsp-kernel: Add cleanup label to fix memory leak in rpmsg_probe
In fastrpc_rpmsg_probe(), 'data' is allocated using kzalloc but not
freed if fastrpc_populate_domain_from_dt() fails. This causes a
potential memory leak.

To fix this, a generic cleanup label 'free_data' is added. On failure
of domain population, the code jumps to this label to free 'data'
before returning.

Change-Id: I25cc2486caab5f901d1d2212b3fba15d93e38255
Signed-off-by: Kumari Pallavi <quic_kpallavi@quicinc.com>